Add DeepLinks, Group View, Group Purchase view, start StoreKit implementation

+++ Fridgy-iOS/Managers/FridgeManager.swift
@@ -9,6 +9,8 @@
 import Foundation
 import CoreData
 
+// In charge of the remote fridge
+// Handles sending updates and taking care of joining/leaving
 class FridgeManager {
     static var shared = FridgeManager()
     
@@ -18,6 +20,8 @@ class FridgeManager {
         // Else check online as may have uninstalled the app
         
     }
+    
+    var inSharedFridge: Bool { UserDefaults.standard.string(forKey: "fridgeId") != nil }
     
     var pendingRecords: [Record] = []
     
@@ -70,7 +74,7 @@ class FridgeManager {
                         localCategory.addToChildren(new)
                     }
                 }
-
+                
                 // Now check that all local items were in the remote and haven't been deleted
                 for localItem in localCategory.items {
                     if !category.items.contains(where: { $0.id == localItem.uniqueId }) {
@@ -87,6 +91,12 @@ class FridgeManager {
             if !fridge.categories.contains(where: {$0.id == category.uniqueId}) {
                 context.delete(category)
             }
+        }
+        
+        UserDefaults.standard.set(fridge.admin, forKey: "admin")
+        
+        if let data = try? JSONEncoder().encode(fridge.users) {
+            UserDefaults.standard.set(data, forKey: "users")
         }
         
         try context.save()
@@ -106,8 +116,9 @@ class FridgeManager {
                 try await NetworkManager.shared.createCategory(id: id, name: name)
             case .deleteCategory(let id):
                 try await NetworkManager.shared.deleteCategory(id: id)
-            case .leaveGroup:
-                try await NetworkManager.shared.leaveFridge()
+            case .leaveGroup(let id):
+                try await NetworkManager.shared.leaveFridge(userId: id)
+                break
         }
         return true
     }
@@ -123,6 +134,7 @@ class FridgeManager {
     }
     
     func addItem(_ item: Item) {
+        guard inSharedFridge else { return }
         let record = Record.addItem(
             name: item.name,
             inShoppingList: item.inShoppingList,
@@ -134,11 +146,13 @@ class FridgeManager {
     }
     
     func deleteItem(id: String, category: String) {
+        guard inSharedFridge else { return }
         let record = Record.deleteItem(id: id, category: category)
         attemptExecuteRecord(record)
     }
     
     func updateItem(_ item: Item) {
+        guard inSharedFridge else { return }
         let record = Record.updateItem(
             name: item.name,
             inShoppingList: item.inShoppingList,
@@ -150,6 +164,7 @@ class FridgeManager {
     }
     
     func changeItemCategory(_ item: Item, previous: String) {
+        guard inSharedFridge else { return }
         let record = Record.changeItemCategory(
             destination: item.category?.uniqueId ?? "",
             id: item.uniqueId,
@@ -159,6 +174,7 @@ class FridgeManager {
     }
     
     func createCategory(_ category: Category) {
+        guard inSharedFridge else { return }
         let record = Record.createCategory(
             name: category.name,
             id: category.uniqueId
@@ -167,20 +183,69 @@ class FridgeManager {
     }
     
     func deleteCategory(id: String) {
+        guard inSharedFridge else { return }
         let record = Record.deleteCategory(
             id: id
         )
         attemptExecuteRecord(record)
     }
     
-    func createFridge(name: String, categories: [Category]) {
+    func createFridge(user: String, name: String, categories: [Category]) async throws {
+        let id = try await NetworkManager.shared.createFridge(name: name, admin: user)
+        try await NetworkManager.shared.joinFridge(userId: user, fridgeId: id)
+        UserDefaults.standard.set(id, forKey: "fridgeId")
+        UserDefaults.standard.set(user, forKey: "admin")
+        let users = try await NetworkManager.shared.getUsers(fridgeId: id)
+        let data = try JSONEncoder().encode(users)
+        UserDefaults.standard.set(data, forKey: "users")
+        
         for category in categories {
             createCategory(category)
             for item in category.items {
                 addItem(item)
             }
         }
-        // Join and rename fridge
+    }
+    
+    func joinFridge(user: String, fridgeId: String, context: NSManagedObjectContext) async throws {
+        try await NetworkManager.shared.joinFridge(userId: user, fridgeId: fridgeId)
+        UserDefaults.standard.set(fridgeId, forKey: "fridgeId")
+        try await fetchFridge(context: context)
+    }
+    
+    func deleteFridge() async throws {
+        if let data = UserDefaults.standard.data(forKey: "users") {
+            let users = try JSONDecoder().decode([User].self, from: data)
+            for user in users {
+                try await NetworkManager.shared.leaveFridge(userId: user.id)
+            }
+        }
+        
+        try await NetworkManager.shared.deleteFridge()
+        UserDefaults.standard.setValue(nil, forKey: "fridgeId")
+        UserDefaults.standard.setValue(nil, forKey: "admin")
+        UserDefaults.standard.setValue(nil, forKey: "users")
+    }
+    
+    func removeUserFromFridge(user: String) async throws {
+        try await NetworkManager.shared.leaveFridge(userId: user)
+        
+        if let data = UserDefaults.standard.data(forKey: "users") {
+            var users = try JSONDecoder().decode([User].self, from: data)
+            
+            users.removeAll(where: { $0.id == user })
+            
+            let newData = try JSONEncoder().encode(users)
+            UserDefaults.standard.set(newData, forKey: "users")
+        }
+    }
+    
+    func leaveFridge(user: String) async throws {
+        
+        try await NetworkManager.shared.leaveFridge(userId: user)
+        UserDefaults.standard.setValue(nil, forKey: "fridgeId")
+        UserDefaults.standard.setValue(nil, forKey: "admin")
+        UserDefaults.standard.setValue(nil, forKey: "users")
     }
 }
 
@@ -192,6 +257,6 @@ enum Record: Codable {
     case changeItemCategory(destination: String, id: String, category: String) // Heavier transaction
     case createCategory(name: String, id: String)
     case deleteCategory(id: String)
-    case leaveGroup
+    case leaveGroup(id: String)
 }
